otdr_monitoring/src/requirements.txt
matplotlib>=3.6.1
numpy>=1.23.2
PySide6-Essentials>=6.3.1
pyusb>=1.2.1
transitions>=0.9
